+++ backend/projects/SmartSocial/src/main/java/de/hsbi/smartsocial/Service/LeaderboardService.java
@@ -35,7 +35,7 @@ public class LeaderboardService {
         return leaderboardRepository.getPersonalStats(id);
     }
 
-    public void addKilometers(Long id, Long kilometers) {
+    public void addKilometers(Long id, double kilometers) {
         Leaderboard leaderboard = leaderboardRepository.getPersonalStats(id);
         leaderboard.setKilometers(leaderboard.getKilometers().add(new java.math.BigDecimal(kilometers)));
         leaderboardRepository.update(leaderboard);
